+++ src/plots/funnel/index.ts
@@ -1,5 +1,9 @@
+import { each } from '@antv/util';
+import { Element } from '@antv/g2';
 import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
+import { StateName, StateCondition, StateObject } from '../..';
+import { getAllElementsRecursively } from '../../utils';
 import { FunnelOptions } from './types';
 import { adaptor } from './adaptor';
 import {
@@ -42,5 +46,39 @@ export class Funnel extends Plot<FunnelOptions> {
    */
   protected getSchemaAdaptor(): Adaptor<FunnelOptions> {
     return adaptor;
+  }
+
+  /**
+   * 设置状态
+   * @param type 状态类型，支持 'active' | 'inactive' | 'selected' 三种
+   * @param conditions 条件，支持数组
+   * @param status 是否激活，默认 true
+   */
+  public setState(type: StateName, condition: StateCondition, status: boolean = true) {
+    const elements = getAllElementsRecursively(this.chart);
+
+    each(elements, (ele: Element) => {
+      if (condition(ele.getData())) {
+        ele.setState(type, status);
+      }
+    });
+  }
+
+  /**
+   * 获取状态
+   */
+  public getStates(): StateObject[] {
+    const elements = getAllElementsRecursively(this.chart);
+
+    const stateObjects: StateObject[] = [];
+    each(elements, (element: Element) => {
+      const data = element.getData();
+      const states = element.getStates();
+      each(states, (state) => {
+        stateObjects.push({ data, state, geometry: element.geometry, element });
+      });
+    });
+
+    return stateObjects;
   }
 }