+++ src/plots/sankey/adaptor.ts
@@ -1,10 +1,35 @@
+import { uniq } from '@antv/util';
 import { interaction, theme } from '../../adaptor/common';
 import { Params } from '../../core/adaptor';
-import { flow } from '../../utils';
+import { deepAssign, flow } from '../../utils';
 import { polygon, edge } from '../../adaptor/geometries';
+import { transformToViewsData } from './helper';
 import { SankeyOptions } from './types';
 import { X_FIELD, Y_FIELD, COLOR_FIELD, EDGES_VIEW_ID, NODES_VIEW_ID } from './constant';
-import { transformToViewsData } from './helper';
+
+/**
+ * 默认配置项 处理
+ * @param params
+ */
+function defaultOptions(params: Params<SankeyOptions>): Params<SankeyOptions> {
+  const { options } = params;
+  const { rawFields = [] } = options;
+
+  return deepAssign(
+    {},
+    {
+      options: {
+        tooltip: {
+          fields: uniq(['name', 'source', 'target', 'value', 'isNode', ...rawFields]),
+        },
+        label: {
+          fields: uniq(['x', 'name', ...rawFields]),
+        },
+      },
+    },
+    params
+  );
+}
 
 /**
  * geometry 处理
@@ -12,7 +37,7 @@ import { transformToViewsData } from './helper';
  */
 function geometry(params: Params<SankeyOptions>): Params<SankeyOptions> {
   const { chart, options } = params;
-  const { color, nodeStyle, edgeStyle, label, tooltip } = options;
+  const { color, nodeStyle, edgeStyle, label, tooltip, nodeState, edgeState } = options;
 
   // 1. 组件，优先设置，因为子 view 会继承配置
   chart.legend(false);
@@ -42,14 +67,7 @@ function geometry(params: Params<SankeyOptions>): Params<SankeyOptions> {
         shape: 'arc',
       },
       tooltip,
-      state: {
-        active: {
-          style: {
-            opacity: 0.8,
-            lineWidth: 0,
-          },
-        },
-      },
+      state: edgeState,
     },
   });
 
@@ -68,6 +86,7 @@ function geometry(params: Params<SankeyOptions>): Params<SankeyOptions> {
       },
       label,
       tooltip,
+      state: nodeState,
     },
   });
 
@@ -135,6 +154,7 @@ export function nodeDraggable(params: Params<SankeyOptions>): Params<SankeyOptio
 export function adaptor(params: Params<SankeyOptions>) {
   // flow 的方式处理所有的配置到 G2 API
   return flow(
+    defaultOptions,
     geometry,
     interaction,
     nodeDraggable,