+++ src/plots/treemap/types.ts
@@ -11,7 +11,10 @@ export interface TreemapOptions extends Omit<Options, 'data'> {
   readonly rectStyle?: StyleAttr;
   /** 层级布局配置 */
   readonly hierarchyConfig?: Omit<HierarchyOption, 'as' | 'type' | 'field'>;
-
+  /**
+   * 附加的 源字段
+   */
+  readonly rawFields?: string[];
   // 矩阵树图 内置一些交互
   /** 下钻交互相关配置 */
   readonly drilldown?: DrillDownCfg;